To fix an execption when a script or other parts are empty.
An exception occurs when a script content or other parts are null. This occurs a rule without checking script contents. For example, if we create a rule like this.
```
    <Rule>
        <Name>Non-standard powershell executable</Name>
        <Conditions>
            <Condition>
                <Property>ApplicationName</Property>
                <Operator>NotContains</Operator>
                <Value>C:\Windows\System32\WindowsPowerShell\v1.0\powershell.exe</Value>
            </Condition>
        </Conditions>
        <Actions>
            <ActionRef>
                <Name>File</Name>
            </ActionRef>
        </Actions>
    </Rule>
```

And it occurs when an application name is WMI and DotNet. And the message is here.
```
Failed to analyze script. Object reference not set to an instance of an object.
   at Engine.ScriptContext.get_Members()
   at Engine.Analyze.Conditions.ListCondition.Analyze(ScriptContext context, Condition condition)
   at Engine.Analyzer.Analyze(ScriptContext scriptContext)
   at Engine.ProtectEngine.Analyze(String script, String path, String app)
```

To avoid this, I added additional checks if the contents are null or not.